Add demo vetting app.
The app is a plain HTML5 web app using the new Stream API
to get a feed from the device camera and the jsqrcode
library to decode QR codes.

It's tested in both Firefox (47.0 on OS X/Android), and
Chrome (51.0.2704.103/81 on OS X/Android).
Chrome requires the page to be served from "secure origins
(such as HTTPS)".